Testing refactored to automatically generate test cases per typeset
- Upgraded linting (black + isort 5)
- Added EmailAddress to CompleteSet
