+++ requirements_dev.txt
@@ -1,7 +1,7 @@
 setuptools>=46.1.3
 wheel>=0.34.2
 black>=19.10b0
-isort>=4.3.21
+isort>=5.0.9
 mypy>=0.770
 recommonmark>=0.6.0
 sphinx_rtd_theme>=0.4.3
