+++ src/visions/application/summaries/summary.py
@@ -3,10 +3,10 @@ from typing import Any, Dict, Type
 import networkx as nx
 import pandas as pd
 
-from visions.application.summaries.frame.dataframe_summary import dataframe_summary
 from visions.application.summaries.frame.dataframe_series_summary import (
     dataframe_series_summary,
 )
+from visions.application.summaries.frame.dataframe_summary import dataframe_summary
 from visions.application.summaries.frame.dataframe_type_summary import (
     dataframe_type_summary,
 )
