+++ src/visions/types/boolean.py
@@ -2,7 +2,7 @@ from typing import Sequence
 
 import numpy as np
 import pandas as pd
-import pandas.api.types as pdt
+from pandas.api import types as pdt
 
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
 from visions.relations.string_to_bool import get_boolean_coercions
@@ -26,7 +26,7 @@ def to_bool(series: pd.Series) -> pd.Series:
 
 
 def _get_relations(cls) -> Sequence[TypeRelation]:
-    from visions.types import Generic, String, Integer, Object
+    from visions.types import Generic, Integer, Object, String
 
     relations = [
         IdentityRelation(cls, Generic),
