+++ src/visions/types/complex.py
@@ -2,7 +2,7 @@ from typing import Sequence
 
 import numpy as np
 import pandas as pd
-import pandas.api.types as pdt
+from pandas.api import types as pdt
 
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
 from visions.types.float import test_is_float
