+++ src/visions/types/date.py
@@ -3,10 +3,10 @@ from typing import Sequence
 
 import pandas as pd
 
-from visions.relations import IdentityRelation, TypeRelation, InferenceRelation
+from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
 from visions.types.type import VisionsBaseType
-from visions.utils.series_utils import nullable_series_contains, class_name_attrs
 from visions.utils.coercion import test_utils
+from visions.utils.series_utils import class_name_attrs, nullable_series_contains
 
 
 def test_date(series):
