+++ src/visions/types/float.py
@@ -2,13 +2,13 @@ from typing import Sequence
 
 import numpy as np
 import pandas as pd
-import pandas.api.types as pdt
+from pandas.api import types as pdt
 
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
 from visions.types.type import VisionsBaseType
 from visions.utils.coercion import test_utils
-from visions.utils.warning_handling import suppress_warnings
 from visions.utils.series_utils import func_nullable_series_contains
+from visions.utils.warning_handling import suppress_warnings
 
 
 def test_string_leading_zeros(series: pd.Series, coerced_series: pd.Series):
@@ -50,7 +50,7 @@ def to_float(series: pd.Series) -> pd.Series:
 
 
 def _get_relations(cls) -> Sequence[TypeRelation]:
-    from visions.types import Generic, String, Complex
+    from visions.types import Complex, Generic, String
 
     relations = [
         IdentityRelation(cls, Generic),
