--- src/visions/types/statistics/__init__.py
@@ -1,1 +1,0 @@
-"""Namespace for statistical types (e.g. nominal, ordinal, interval)"""
