+++ src/visions/types/string.py
@@ -1,7 +1,7 @@
 from typing import Sequence
 
 import pandas as pd
-import pandas.api.types as pdt
+from pandas.api import types as pdt
 
 from visions.relations import IdentityRelation, TypeRelation
 from visions.types.type import VisionsBaseType
