+++ src/visions/utils/__init__.py
@@ -1,12 +1,12 @@
 """ Utilities suite for visions """
 from visions.utils.coercion import test_utils
-from visions.utils.profiling import profile_type
 
 # from visions.utils.images import image_utils
 from visions.utils.monkeypatches import imghdr_patch, pathlib_patch
-from visions.utils.warning_handling import suppress_warnings
+from visions.utils.profiling import profile_type
 from visions.utils.series_utils import (
-    nullable_series_contains,
     func_nullable_series_contains,
     isinstance_attrs,
+    nullable_series_contains,
 )
+from visions.utils.warning_handling import suppress_warnings
