+++ tests/test_type_inference.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.series import get_series, infer_series_type_map
+from tests.series import get_series
 from visions import EmailAddress
 from visions.types import Generic
 from visions.typesets import CompleteSet
@@ -18,27 +18,6 @@ def pytest_generate_tests(metafunc):
             argsvalues.append(pytest.param(series, **args))
 
         metafunc.parametrize(argnames=["series"], argvalues=argsvalues)
-    if metafunc.function.__name__ == "test_inference":
-        argsvalues = []
-        inferred_series_type_map = infer_series_type_map()
-        for series in _test_suite:
-            expected_type = inferred_series_type_map[series.name]
-            for test_type in typeset.types:
-                args = {
-                    "id": "{name} x {type} expected {expected}".format(
-                        name=series.name,
-                        type=test_type,
-                        expected=test_type == expected_type,
-                    )
-                }
-                difference = test_type != expected_type
-                argsvalues.append(
-                    pytest.param(series, test_type, typeset, difference, **args)
-                )
-        metafunc.parametrize(
-            argnames=["series", "expected_type", "typeset", "difference"],
-            argvalues=argsvalues,
-        )
 
 
 def test_consistency(series):
@@ -65,8 +44,3 @@ def _traverse_relation_graph(series, G, node=Generic):
 # What does this test? It doesn't explicitly invoke the actual traversal code.
 def test_traversal_mutex(series):
     _traverse_relation_graph(series, typeset.relation_graph)
-
-
-def test_inference(series, expected_type, typeset, difference):
-    inferred_type = typeset.infer_series_type(series)
-    assert (inferred_type == expected_type) != difference
